requirements.txt
asgiref==3.5.0
boto==2.49.0
boto3==1.21.13
botocore==1.24.13
Django==3.2.12
django-cors-headers==3.11.0
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
jmespath==0.10.0
Pillow==9.0.1
pkg-resources==0.0.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.8
uWSGI==2.0.20
